+++ R/namespace.R
@@ -1,4 +1,4 @@
-#' @import ggplot2 chk
+#' @import ggplot2 chk yesno
 #' @importFrom stats setNames
 #' @importFrom graphics plot
 NULL